try to minimize lag
