+++ project/pong/consumers.py
@@ -5,75 +5,101 @@ from django.shortcuts import get_object_or_404
 from channels.generic.websocket import AsyncWebsocketConsumer
 from channels.layers import get_channel_layer
 from asgiref.sync import sync_to_async
-from database.models import Game as GameDB, Player as PlayerDB
 
 games_pool = {}
+# logger = logging.getLogger(__name__)
+physics_lock = asyncio.Lock()
 
-class MatchmakingPongConsumer(AsyncWebsocketConsumer):
+class LocalPongConsumer(AsyncWebsocketConsumer):
     async def connect(self):
+        from database.models import Game as GameDB, Player as PlayerDB
         from .game_objects import Game, Player
 
         self.game_id = self.scope["url_route"]["kwargs"]["game_id"]
         self.group_name = f"game_{self.game_id}"
 
-        # Ajouter le joueur à un groupe de websockets
         await self.channel_layer.group_add(self.group_name, self.channel_name)
 
         # Vérifier si le jeu est déjà en mémoire
-        if self.game_id in games_pool:
-            logger.info(f"Connecting to an existing game instance for game {self.game_id}")
-            games_pool[self.game_id]["connections"] += 1
-            self.game = games_pool[self.game_id]["game"]
-            self.log_game_state("Existing Game param")
-        else:
-            # Essayer de récupérer la game depuis la base de données (matchmaking)
-            GameDB = await sync_to_async(self.get_game_from_db)(self.game_id)
-
-            if GameDB is not None:
-                logger.info(f"Game found in the database for game {self.game_id}")
-                # Récupérer les joueurs depuis la base de données
-                player1 = await sync_to_async(lambda: GameDB.player1)()
-                player2 = await sync_to_async(lambda: GameDB.player2)()
-
-                # Créer une nouvelle instance de jeu avec les joueurs de la base de données
-                self.game = Game(self.game_id, [], 2, 1280, 720, self.notifyEvent, "dbGame")
-                self.game.addPlayer(Player(id=player1.id, skin="skin1", name=player1.name), 0)
-                self.game.addPlayer(Player(id=player2.id, skin="skin2", name=player2.name), 1)
+        async with physics_lock:
+            if self.game_id in games_pool:
+                logger.info(
+                    f"Connecting to an existing game instance for game {self.game_id}"
+                )
+                games_pool[self.game_id]["connections"] += 1
+                self.game = games_pool[self.game_id]["game"]
+                self.log_game_state("Existing Game param")
             else:
-                logger.info(f"No game in the database, creating a new matchmaking game instance for game {self.game_id}")
-                # Créer un nouveau jeu pour le matchmaking si aucun jeu DB n'existe
-                self.game = Game(self.game_id, [], 2, 1280, 720, self.notifyEvent, "matchmakingGame")
-                player1 = Player(id=0, skin="skin1", name="Player1")
-                player2 = Player(id=1, skin="skin2", name="Player2")
-                self.game.addPlayer(player1, 0)  # Joueur 1 (gauche)
-                self.game.addPlayer(player2, 1)  # Joueur 2 (droite)
+                # Essayer de récupérer la game depuis la base de données
+                GameDB = await sync_to_async(self.get_game_from_db)(self.game_id)
 
-            # Démarrer la physique du jeu
-            if not hasattr(self, "physics_task") or self.physics_task.done():
-                self.physics_task = asyncio.create_task(self.game.physics_loop())
-                self.game.running = True
+                if GameDB is not None:
+                    logger.info(f"Game found in the database for game {self.game_id}")
+                    # Récupérer les joueurs depuis la base de données
+                    player1 = await sync_to_async(lambda: GameDB.player1)()
+                    player2 = await sync_to_async(lambda: GameDB.player2)()
 
-            # Ajouter le jeu au pool de jeux actifs
-            games_pool[self.game_id] = {
-                "game": self.game,
-                "connections": 1,  # Premier client connecté
-            }
+                    # Créer une nouvelle instance de game avec les joueurs de la base de données
+                    self.game = Game(
+                        self.game_id, [], 2, 1280, 720, self.notifyEvent, "dbGame"
+                    )
+                    self.game.addPlayer(
+                        Player(id=player1.id, skin="skin1", name=player1.name), 0
+                    )
+                    self.game.addPlayer(
+                        Player(id=player2.id, skin="skin2", name=player2.name), 1
+                    )
+                else:
+                    logger.info(
+                        f"No game in the database, creating a new local game instance for game {self.game_id}"
+                    )
+                    # Créer un jeu local si aucun jeu DB n'existe
+                    self.game = Game(
+                        self.game_id, [], 2, 1280, 720, self.notifyEvent, "localGame"
+                    )
+                    player1 = Player(id=0, skin="skin1", name="Player1")
+                    player2 = Player(id=1, skin="skin2", name="Player2")
+                    self.game.addPlayer(player1, 0)  # Joueur 1 (gauche)
+                    self.game.addPlayer(player2, 1)  # Joueur 2 (droite)
 
-        # Accepter la connexion WebSocket
+                # Démarrer la physique
+                if not hasattr(self, "physics_task") or self.physics_task.done():
+                    self.physics_task = asyncio.create_task(self.game.physics_loop())
+                    self.game.running = True
+
+                # Ajouter le jeu à games_pool
+                games_pool[self.game_id] = {
+                    "game": self.game,
+                    "connections": 1,  # Premier client connecté
+                }
+
         await self.accept()
 
         # Envoyer l'état initial ou mis à jour du jeu
-        await self.send(text_data=json.dumps({"type": "init", "game": await build_game_state(self.game)}))
+        await self.send(
+            text_data=json.dumps(
+                {"type": "init", "game": await build_game_state(self.game)}
+            )
+        )
 
     def get_game_from_db(self, game_id):
-        """Récupérer une instance de Game depuis la base de données"""
+        """
+        Méthode pour récupérer une instance de Game depuis la base de données
+        """
+        from database.models import Game as GameDB, Player as PlayerDB
+
         try:
             return GameDB.objects.get(game_ws_id=game_id)
         except GameDB.DoesNotExist:
             return None
 
     def log_game_state(self, context):
-        """Log du statut du jeu avec les positions et états actuels"""
+        """
+        Log the current state of the game including paddle positions, ball position,
+        ball speed, and scores.
+
+        :param context: A string to describe the context from which this log is called.
+        """
         paddle_positions = [(paddle.x, paddle.y) for paddle in self.game.paddles]
         ball_position = (self.game.ball.x, self.game.ball.y)
         ball_speed = self.game.ball.speed
@@ -88,20 +114,17 @@ class MatchmakingPongConsumer(AsyncWebsocketConsumer):
         )
 
     async def notifyEvent(self, eventData):
-        """Notifier tous les joueurs des événements du jeu via WebSocket"""
         await self.channel_layer.group_send(
             self.group_name, {"type": "gameEvent", "event": eventData}
         )
 
     async def gameEvent(self, event):
-        """Envoyer un événement de jeu à ce client spécifique"""
+        # Envoyer l'événement reçu à ce client spécifique
         await self.send(text_data=json.dumps(event["event"]))
 
     async def receive(self, text_data=None, bytes_data=None):
-        """Gérer les entrées des joueurs (clavier)"""
         text_data_json = json.loads(text_data)
 
-        # Identifier quel joueur a envoyé l'input
         if text_data_json["key"] in ["w", "s", "space"]:
             who = 0  # Joueur 1 (gauche)
         elif text_data_json["key"] in ["arrowup", "arrowdown"]:
@@ -109,48 +132,47 @@ class MatchmakingPongConsumer(AsyncWebsocketConsumer):
         else:
             return
 
-        # Gérer l'appui et le relâchement des touches
         if text_data_json["type"] == "keydown" or text_data_json["type"] == "keyup":
-            await handle_key(self.game, text_data_json["type"], text_data_json["key"], who)
+            await handle_key(
+                self.game, text_data_json["type"], text_data_json["key"], who
+            )
 
     async def game_update(self):
-        """Mettre à jour l'état du jeu à chaque frame (60 FPS)"""
         while self.game.running:
             await asyncio.sleep(1 / 60)
             await self.send(text_data=json.dumps(await update_game_state(self.game)))
 
     async def disconnect(self, close_code):
-        """Gérer la déconnexion des clients"""
         logger.info(f"Client disconnected from game {self.game_id}")
 
-        # Retirer le client du groupe WebSocket
+        # Retirer le client du groupe
         await self.channel_layer.group_discard(self.group_name, self.channel_name)
 
-        # Mise à jour du pool de jeux et gestion des déconnexions
         if self.game_id in games_pool:
             games_pool[self.game_id]["connections"] -= 1
 
             # Attendre un délai pour permettre aux clients de se reconnecter
             await asyncio.sleep(10)
 
-            # Nettoyer si aucun joueur n'est connecté
-            if games_pool[self.game_id]["connections"] <= 0:
-                logger.info(f"No more connections for game {self.game_id}. Cleaning up.")
+            # Vérifier s'il reste des connexions après le délai
+        if self.game_id in games_pool and games_pool[self.game_id]["connections"] <= 0:
+            logger.info(f"No more connections for game {self.game_id}. Cleaning up.")
 
-                # Annuler la tâche de physique si elle est toujours en cours
-                if hasattr(self, "physics_task") and not self.physics_task.done():
-                    self.physics_task.cancel()
-                    try:
-                        await self.physics_task
-                    except asyncio.CancelledError:
-                        logger.info(f"Physics task for game {self.game_id} has been cancelled.")
+            # Annuler la tâche de physique si elle est en cours d'exécution
+            if hasattr(self, "physics_task") and not self.physics_task.done():
+                self.physics_task.cancel()
+                try:
+                    await self.physics_task
+                except asyncio.CancelledError:
+                    logger.info(
+                        f"Physics task for game {self.game_id} has been cancelled."
+                    )
 
-                # Supprimer le jeu du pool
-                games_pool.pop(self.game_id, None)
-                logger.info(f"Current games_pool state after cleanup: {games_pool}")
+            games_pool.pop(self.game_id, None)
+            logger.info(f"Current games_pool state after cleanup: {games_pool}")
+
 
 async def handle_key(game, types, key, who):
-    """Gérer l'entrée clavier des joueurs"""
     if types not in ["keydown", "keyup"]:
         return
 
@@ -176,20 +198,27 @@ async def handle_key(game, types, key, who):
 
     paddle = game.paddles[who]
 
+    # Mettre à jour l'état de la touche dans le dictionnaire
     if types == "keydown":
         paddle.keys_pressed[action] = True
     elif types == "keyup":
         paddle.keys_pressed[action] = False
 
+    # Gestion des mouvements en fonction des touches pressées
     if paddle.keys_pressed["up"] and not paddle.keys_pressed["down"]:
         paddle.velocity = -1  # Monter
     elif paddle.keys_pressed["down"] and not paddle.keys_pressed["up"]:
         paddle.velocity = 1  # Descendre
     else:
-        paddle.velocity = 0  # Aucun mouvement si les deux touches sont relâchées ou pressées
+        paddle.velocity = (
+            0  # Aucun mouvement si les deux touches sont relâchées ou pressées
+        )
+    # logger.debug(f"Paddle update - Player: {who}, Action: {action}, Type: {types}, New Velocity: {game.paddles[who].velocity}")
+    # logger.debug(f"New Paddle0 Position: x={game.paddles[0].x} y={game.paddles[0].y}")
+    # logger.debug(f"New Paddle1 Position: x={game.paddles[1].x} y={game.paddles[1].y}")
+
 
 async def build_game_state(game):
-    """Créer l'état du jeu pour initialiser la partie"""
     return {
         "type": "initGame",
         "game_id": game.id,
@@ -222,8 +251,8 @@ async def build_game_state(game):
         "isPaused": game.isPaused,
     }
 
+
 async def update_game_state(game):
-    """Mettre à jour l'état du jeu à chaque frame"""
     return {
         "type": "update",
         "game_id": game.id,
